fix(structure): Move channel related layout components to a folder
move all channel related components from shared to its own folder
