+++ components/shared/Channel/channel-card.tsx
